Move ClickToCopy to shared components
