+++ src/components/ClickToCopy.tsx
